avant de tout casser, je gere les arrondis

+++ srcs/tools_f_2.c
@@ -6,7 +6,7 @@
 /*   By: edbaudou <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/05 13:01:54 by edbaudou          #+#    #+#             */
-/*   Updated: 2019/09/08 16:18:07 by edbaudou         ###   ########.fr       */
+/*   Updated: 2019/09/14 16:58:23 by edbaudou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,9 +19,9 @@ long double		ft_get_arg_f(t_printf *v_printf)
 	if (v_printf->flags & LF)
 		return ((long double)va_arg(v_printf->ap, long double));
 	else if (v_printf->flags & L)
-		return ((double)va_arg(v_printf->ap, double));
+		return ((long double)va_arg(v_printf->ap, double));
 	else 
-		return ((float)va_arg(v_printf->ap, double));
+		return ((long double)va_arg(v_printf->ap, double));
 }
 
 void			ft_get_mant(t_float *f)
